[lint] decompose JSDoc analysis and preserve example siblings
